+++ perception/radar_fusion_to_detected_object/include/radar_fusion_to_detected_object/radar_fusion_to_detected_object_node.hpp
@@ -19,7 +19,7 @@
 #include "message_filters/subscriber.h"
 #include "message_filters/sync_policies/approximate_time.h"
 #include "message_filters/synchronizer.h"
-#include "radar_fusion_to_detected_object.hpp"
+#include "radar_fusion_to_detected_object/radar_fusion_to_detected_object.hpp"
 #include "rclcpp/rclcpp.hpp"
 
 #include "autoware_auto_perception_msgs/msg/detected_objects.hpp"
